core/url: Clarify the copyright coverage
